requirements.txt
playwright>=1.44.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
boto3>=1.34.0
